refactor: 채팅 오류 클라이언트 측에 보여주기
