+++ domain/src/main/java/semicolon/viewtist/global/exception/GlobalExceptionHandler.java
@@ -3,6 +3,7 @@ package semicolon.viewtist.global.exception;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import semicolon.viewtist.chatting.exception.ChattingException;
 import semicolon.viewtist.jwt.exception.JwtException;
 import semicolon.viewtist.liveStreaming.exception.LiveStreamingException;
 import semicolon.viewtist.user.exception.UserException;
@@ -15,7 +16,10 @@ public class GlobalExceptionHandler {
   public ResponseEntity<?> handlerLiveStreamingException(LiveStreamingException e) {
     return toResponse(e.getErrorCode(), e.getMessage());
   }
-
+  @ExceptionHandler(ChattingException.class)
+  public ResponseEntity<?> handlerChattingException(ChattingException e) {
+    return toResponse(e.getErrorCode(), e.getMessage());
+  }
   @ExceptionHandler(UserException.class)
   public ResponseEntity<?> handlerUserException(UserException e) {
     return toResponse(e.getErrorCode(), e.getMessage());
